streamlit-snowflake-app/requirements.txt
streamlit==1.31.0
pandas==2.1.4
snowflake-connector-python==3.6.0
plotly==5.18.0